Update template.mdL5 Removed Duplicate Space L8-13 List in BlockQuote L15 19 31 Colon Added - clarity L17 Corrected incorrect use of conjunctive clauses L21 Added to the URL - clarity L27 L43 - Mistake proofing language around screenshots L28-29 34-35 44-45 Placed Screenshots in Blockquote
L5 - Line contained addition spacing in paragraph between "And" and "Horizontally.  

L8-13 - List items in blockquotes adds visual clarity to the presentation of items and reduces information signal confusion, in both Tracker and Report. This item must be fixed on every instance of this item that I find, having this change in Markdown template will reduce the # of cycles spent editing findings in tracker. 

L15 19 31 - Bolded subsection headings with a comma reduce the time taken to convert to report format.

L17 - Paragraph incorrectly joins two clauses. One solution is to add the specified comma. 

L21 - adding "to the URL" to this line aids readability for the non-technical audience, which is a concern for reporting since our items will be accessible by non-technical administrative staff at any client's office/workspace. 

L27 43 Screenshot(s) et al - adding this language reduces the time required to ensure that language surrounding screenshot usage is in a correct subject verb agreement.  some researchers, especially non-native english speaking global researchers, do not excel intuitively at the english language like most native english speakers.  Rather than provide additional training to every researcher, or pay for additional editing hours, fix the problem at an administrative level with a language that is error-proof in all usages and in line with professional writing standards worldwide.  

L 28-29 34-35 44-45 - Screenshots in blockqoute adds clarity to the communication of information in tracker.  Screenshots in blockqoute also enable easier editing in Report.  In tracker, blockqoute clearly delineates that a screenshot is attached to, or a sub item, of the step it is listed under.  in report, the blockqoute separates the items from the 0 indent and enables programmatic adjustment to photo items in editor without compromising the integrity of the document's formatting.