Update .markdownlint.json
`"MD041": false,` tells the linting tool to ignore the rule that says that the first line must start with a heading. e.g. # Title of File.

This will allow the template.md files to start with the description of the vulnerability. A feature which will decrease the time it takes  for penetration testing report creation.